+++ src/main/java/com/example/movie/movie/repository/MovieRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -16,11 +17,11 @@ public interface MovieRepository extends CrudRepository<Movie, Long> {
 
 
     @Query("SELECT m FROM Movie m WHERE m.purchasePrice BETWEEN ?1 AND ?2")
-    public List<Movie> filterByPurchasePrice(Long minPrice, Long maxPrice);
+    public List<Movie> filterByPurchasePrice(BigDecimal minPrice, BigDecimal maxPrice);
 
 
     @Query("SELECT m FROM Movie m WHERE m.rentalPrice BETWEEN ?1 AND ?2")
-    public List<Movie> filterByRentalPrice(Long minPrice, Long maxPrice);
+    public List<Movie> filterByRentalPrice(BigDecimal minPrice, BigDecimal maxPrice);
 
 //Спросить CALL
     @Query(value = "{CALL SP_get_movies_by_genre()}", nativeQuery = true)
